pyproject.toml
[tool.poetry]
name = "llm-intent-orchestration"
version = "0.1.0"
description = "Simple LLM for medical user intent in biomedical environment"
authors = ["Huyen-Trang Le", "Nguyen-Huong-Giang Le", "Massinissa Hamidi"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10.0,<3.13"
crewai = ">=0.40.0,<1.0.0"
langchain-google-genai = ">=1.0.0,<2.0.0"
google-generativeai = ">=0.5.0"
unstructured = '==0.10.25'
pyowm = '3.3.0'
tools = "^0.1.9"
python-dotenv = ">=1.1.0"
uvicorn = ">=0.35"
pydantic = "^2.0.0"
influxdb-client = "^1.18.0"
fastapi = "^0.109.0"


# syntax error check 
fastmcp = "^3.2.3"
[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

# styling issue check 
[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

# python tool management depencies
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"